PriceHistory completed and costHistory

+++ src/main/java/org/jog/crudinventory/persistence/repository/CostHistoryRepository.java
@@ -3,5 +3,5 @@ package org.jog.crudinventory.persistence.repository;
 import org.jog.crudinventory.persistence.entity.CostHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CostHistoryRepository extends JpaRepository<CostHistory, Long> {
+public interface CostHistoryRepository extends JpaRepository<CostHistory, Integer> {
 }
